refactor(web): one donut, one row shape
`portfolio/Overview.jsx` carried a hand-copied duplicate of the shared
`Donut`, with its own 7-colour palette. Every later chart change would
land twice across four call sites, so merging makes the responsive diff
smaller rather than larger. The merged component lives beside `api.js`
because both sections draw it.

The copies were not identical, so this is a small behavioural merge
rather than a lift-and-delete. The 8-entry palette wins — the two only
differ once a donut has 8+ slices, where the 7-entry one wraps back to
blue. Only the copy sorted, and it sorted the caller's array in place
during render; the merged component sorts a copy, descending, with no
`sort` prop, because `portfolio/spending.py:58` is `ORDER BY v DESC` and
the spending call sites already depend on arriving that way.

The legend beneath becomes one row shape at every width: a full-width
track at `width: {pct}%` with the fill behind the text, a colour chip,
and the name as a growing flex child. That deletes both the 220px bar
constant — which stopped being proportional the moment a card was wider
than 220px — and the fixed 130px name column, which reserved space it
often did not need. One code path rather than a two-shape conditional in
two component files.

Refs #25
